axcpu::init

Function init_trap

pub fn init_trap()

Initializes trap handling on the current CPU.

In detail, it initializes the GDT, IDT on x86_64 platforms (init_gdt and init_idt). If the uspace feature is enabled, it also initializes relevant model-specific registers to configure the handler for syscall instruction (init_syscall).

Notes

Before calling this function, the initialization function of the percpu crate should have been invoked to ensure that the per-CPU data structures are set up correctly.